Refactored to make calculating contact since a date and last contact added time generic so it can be shared by interactions.
